Use consumer proguard
